Add classes for create_person, create_book and create_rental
