+++ Create/create.rb
@@ -1,0 +1,13 @@
+class Create
+  def create_book
+    raise NotImplementedError, 'You should implement this method'
+  end
+
+  def create_person
+    raise NotImplementedError, 'You should implement this method'
+  end
+
+  def create_rental
+    raise NotImplementedError, 'You should implement this method'
+  end
+end
