--- create.rb
@@ -1,5 +1,0 @@
-class Create
-  def create
-    raise NotImplementedError, 'You should implement this method'
-  end
-end
